+++ src/models/Settings.php
@@ -33,7 +33,7 @@ class Settings extends Model
         'cardlets' => ['width' => 150, 'height' => 150, 'format' => 'webp'],
         'cards' => ['width' => 400, 'height' => 200, 'format' => 'webp'],
         'line' => null, // no image in line layout
-        'table' => ['width' => 35, 'height' => 35, 'format' => 'webp']
+        'table' => ['width' => 50, 'height' => 50, 'format' => 'webp']
     ];
     public string $useCSS = 'all';
     public array $layoutSizes = [