+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * The `getPages` method now returns a collection of `Page` models and is moved to `ContenoverviewService`.
 * Added `query` section setting.
 * Added `imageRatio` section setting.
+* Added `fallbackImageField` section config setting.
 * CSS now uses container queries.
 * * Added a `containerBreakpointColumns` plugin setting that defines container breakpoints and grid columns for layouts. 
 * * A new (temporary) plugin setting `useCSS` can be set to `all` to load the old CSS for legacy browsers, or `modern` to load a polyfill.
@@ -14,7 +15,8 @@
 * Added `titleObjectTemplate` section config setting (defaults to `{title}`)
 * Added `table` layout
 * Added `size` section config setting.
-* Added `fallbackImageField` section config setting.
+* Added `DefineImageEvent`
+
 
 
 ### Enhanced customization